+++ src/mofgbmlpy/gbml/objectives/pittsburgh/num_rules.pyx
@@ -5,13 +5,13 @@ from mofgbmlpy.gbml.solution.pittsburgh_solution cimport PittsburghSolution
 
 cdef class NumRules(ObjectiveFunction):
     """Objective function that considers the number of rules (Michigan solutions) to evaluate Pittsburgh solutions"""
-    cpdef void run(self, AbstractSolution[:] solutions, int obj_index, float[:] out):
+    cpdef void run(self, AbstractSolution[:] solutions, int obj_index, double[:] out):
         """Run the objective function on the given parameters
 
         Args:
             solutions (PittsburghSolution[]): Solutions that are evaluated
             obj_index (int): Index of the objective in the solution objectives array
-            out (float[]): Output array, it will contain the objective value of all the solutions
+            out (double[]): Output array, it will contain the objective value of all the solutions
         """
         cdef int i = 0
         cdef PittsburghSolution sol
