+++ src/mofgbmlpy/utility/lru_cache.pxd
@@ -8,7 +8,7 @@ cdef extern from "lru_cache.h":
         LRUCache()
         size_t combine_keys(int key1, int key2);
         bint has(int key1, int key2)
-        float get(int key1, int key2)
-        void put(int key1, int key2, float value)
+        double get(int key1, int key2)
+        void put(int key1, int key2, double value)
         int get_max_size()
         int get_size()
